+++ src/ByteDev.Hibp/HibpClient.cs
@@ -15,6 +15,7 @@ namespace ByteDev.Hibp
     public class HibpClient : IHibpClient
     {
         private readonly HttpClient _httpClient;
+        private readonly HibpClientOptions _options;
         private readonly HttpRequestMessageFactory _requestFactory;
 
         /// <summary>
@@ -23,9 +24,21 @@ namespace ByteDev.Hibp
         /// <param name="httpClient">HttpClient to use in all requests to the API.</param>
         /// <param name="apiKey">Authorization key for the API.</param>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="httpClient" /> is null.</exception>
-        public HibpClient(HttpClient httpClient, string apiKey)
+        public HibpClient(HttpClient httpClient, string apiKey) : this (httpClient, apiKey, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="T:ByteDev.Hibp.HibpClient" /> class.
+        /// </summary>
+        /// <param name="httpClient">HttpClient to use in all requests to the API.</param>
+        /// <param name="apiKey">Authorization key for the API.</param>
+        /// <param name="options">Additional settings for the HIBP client.</param>
+        /// <exception cref="T:System.ArgumentNullException"><paramref name="httpClient" /> is null.</exception>
+        public HibpClient(HttpClient httpClient, string apiKey, HibpClientOptions options)
         {
             _httpClient = httpClient ?? throw new ArgumentNullException(nameof(httpClient));
+            _options = options ?? new HibpClientOptions();
 
             _requestFactory = new HttpRequestMessageFactory(apiKey);
         }
@@ -43,9 +56,7 @@ namespace ByteDev.Hibp
         {
             var uri = HibpUriFactory.CreateBreachedAccountUri(emailAddress, options);
 
-            var request = _requestFactory.CreateGet(uri);
-
-            var response = await _httpClient.SendAsync(request, cancellationToken);
+            var response = await ApiGetAsync(uri, cancellationToken);
 
             return await HibpResponseFactory.CreateBreachResponsesAsync(response);
         }
@@ -61,9 +72,7 @@ namespace ByteDev.Hibp
         {
             var uri = HibpUriFactory.CreateBreachedSiteUri(domain);
 
-            var request = _requestFactory.CreateGet(uri);
-
-            var response = await _httpClient.SendAsync(request, cancellationToken);
+            var response = await ApiGetAsync(uri, cancellationToken);
 
             return await HibpResponseFactory.CreateBreachResponsesAsync(response);
         }
@@ -80,9 +89,7 @@ namespace ByteDev.Hibp
         {
             var uri = HibpUriFactory.CreateBreachSiteByNameUri(breachName);
 
-            var request = _requestFactory.CreateGet(uri);
-
-            var response = await _httpClient.SendAsync(request, cancellationToken);
+            var response = await ApiGetAsync(uri, cancellationToken);
 
             return await HibpResponseFactory.CreateBreachResponseAsync(response);
         }
@@ -97,9 +104,7 @@ namespace ByteDev.Hibp
         {
             var uri = HibpUriFactory.CreateDataClassesUri();
 
-            var request = _requestFactory.CreateGet(uri);
-
-            var response = await _httpClient.SendAsync(request, cancellationToken);
+            var response = await ApiGetAsync(uri, cancellationToken);
 
             return await HibpResponseFactory.CreateDataClassesAsync(response);
         }
@@ -116,11 +121,27 @@ namespace ByteDev.Hibp
         {
             var uri = HibpUriFactory.CreateAccountPastesUri(emailAddress);
 
-            var request = _requestFactory.CreateGet(uri);
-
-            var response = await _httpClient.SendAsync(request, cancellationToken);
+            var response = await ApiGetAsync(uri, cancellationToken);
 
             return await HibpResponseFactory.CreatePasteResponsesAsync(response);
+        }
+
+        private async Task<HttpResponseMessage> ApiGetAsync(Uri uri, CancellationToken cancellationToken)
+        {
+            var response = await _httpClient.SendAsync(_requestFactory.CreateGet(uri), cancellationToken);
+
+            if (_options.RetryOnRateLimitExceeded && response.IsRateLimitedExceeded())
+            {
+                var rateLimitResponse = await response.DeserializeAsync<ApiRateLimitExceededResponse>();
+
+                var seconds = rateLimitResponse.GetRetrySeconds();
+                
+                await Task.Delay(TimeSpan.FromSeconds(seconds), cancellationToken);
+
+                response = await _httpClient.SendAsync(_requestFactory.CreateGet(uri), cancellationToken);
+            }
+
+            return response;
         }
     }
 }