+++ tests/ByteDev.Hibp.IntTests/EmailAddresses.cs
@@ -1,0 +1,9 @@
+﻿namespace ByteDev.Hibp.IntTests
+{
+    public static class EmailAddresses
+    {
+        public const string Pwned = @"johnsmith@gmail.com";
+
+        public const string NotPwned = @"djs834sskldi4999dspo@gmail.com";
+    }
+}